Refactor: Added notes, added solution with no additional data structures and modified section name
